third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryServices/services-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(function() {
  return setBluetoothFakeAdapter('TwoHeartRateServicesAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['generic_access']}))
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryServices())
    .then(services => {
      assert_equals(services.length, 3);
      assert_equals(services[0].uuid,
                    BluetoothUUID.getService('generic_access'));
      assert_equals(services[1].uuid,
                    BluetoothUUID.getService('heart_rate'));
      assert_equals(services[2].uuid,
                    BluetoothUUID.getService('heart_rate'));
      assert_true(services[0].isPrimary);
      assert_true(services[1].isPrimary);
      assert_true(services[2].isPrimary);
    });
}, 'Find all services in a device.');
</script>
